Struct capabilities::Capabilities

pub struct Capabilities {
    // some fields omitted
}

A capability set that can be manipulated.

Methods

impl Capabilities

fn new() -> Option<Capabilities>

Create a new empty capability set

fn from_fd(fd: isize) -> Option<Capabilities>

Create a capability set from the specified file descriptor

fn from_file(path: &str) -> Option<Capabilities>

Create a capability set base on the supplied file path

fn from_pid(pid: isize) -> Option<Capabilities>

Create a capability set from the supplied process ID.

fn from_current_proc() -> Option<Capabilities>

Create a capability set based on the current processes capabilities.

fn reset_all(&mut self)

Clear all the entries in the capability set.

fn reset_flag(&mut self, flag: Flag)

Clear all instances of the supplied flag.

fn check(&self, cap: Capability, flag: Flag) -> bool

Check if the supplied capability has the flag set in this capability set.

fn update(&mut self, caps: &[Capability], flag: Flag, set: bool) -> bool

Update the capability set adding the supplied capabilities. Each of the supplied capabilities will have the flag set or cleared depending on the value supplied for set.

fn apply(&self) -> Result<(), Error>

Attempt to apply the capability set to the current process.

fn apply_to_fd(&self, fd: i32) -> Result<(), Error>

Attempt to apply the capability set to the supplied file descriptor.

fn apply_to_file(&self, path: &str) -> Result<(), Error>

Attempt to apply the capability set to the supplied file.